+++ app/src/main/java/de/baumann/browser/view/NinjaWebView.java
@@ -455,6 +455,13 @@ public class NinjaWebView extends WebView implements AlbumController {
     }
 
     @Override
+    protected void onWindowVisibilityChanged(int visibility) {
+        if (sp.getBoolean("sp_audioBackground", false)) {
+            if (visibility != View.GONE) super.onWindowVisibilityChanged(View.VISIBLE);
+        }
+    }
+
+    @Override
     public synchronized void stopLoading(){
         stopped=true;
         super.stopLoading();